backend/triage-service/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# OpenAI Agent SDK
openai==1.12.0

# Dapr SDK
dapr-sdk==1.12.0

# Security & Auth
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
pyyaml==6.0.1
psutil==5.9.6

# Kafka
confluent-kafka==2.3.0

# JWT enhanced
cryptography==41.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Performance monitoring
prometheus-client==0.19.0